Fix e2e for AvaTax
