+++ apps/avatax/e2e/tests/checkout-basic-product-without-tax-code.spec.ts
@@ -7,7 +7,7 @@ import {
   CompleteCheckout,
   CreateCheckout,
 } from "../generated/graphql";
-import { getCompleteMoney } from "../utils/moneyUtils";
+import { getCompleteMoney } from "../utils/money";
 
 // Testmo: https://saleor.testmo.net/repositories/6?case_id=16233
 describe("App should calculate taxes for checkout with product without tax class TC: AVATAX_1", () => {
