+++ apps/avatax/e2e/tests/draft-order-with-order-promotion-subtotal.spec.ts
@@ -8,40 +8,43 @@ import {
   DraftOrderComplete,
   DraftOrderUpdateAddress,
   DraftOrderUpdateShippingMethod,
-  OrderDiscountAdd,
+  OrderLineUpdate,
   StaffUserTokenCreate,
 } from "../generated/graphql";
-import { getCompleteMoney } from "../utils/moneyUtils";
+import { getCompleteMoney, getMoney } from "../utils/money";
 
-// Testmo: https://saleor.testmo.net/repositories/6?group_id=139&case_id=18386
-describe("App should calculate taxes for draft order with manual total discount applied TC: AVATAX_22", () => {
-  const testCase = e2e("Product with tax class [pricesEnteredWithTax: True]");
+// Testmo: https://saleor.testmo.net/repositories/6?group_id=4846&case_id=24175
+describe("App should calculate taxes on draft order when order promotion is applied TC: AVATAX_32", () => {
+  const testCase = e2e("Draft order with order promotion, [pricesEnteredWithTax: False]");
   const staffCredentials = {
     email: env.E2E_USER_NAME,
     password: env.E2E_USER_PASSWORD,
   };
 
   const CURRENCY = "USD";
+  const TOTAL_GROSS_PRICE_BEFORE_SHIPPING = 326.08;
+  const TOTAL_NET_PRICE_BEFORE_SHIPPING = 299.5;
+  const TOTAL_TAX_PRICE_BEFORE_SHIPPING = 26.58;
 
-  const TOTAL_GROSS_PRICE_BEFORE_SHIPPING = 15;
-  const TOTAL_NET_PRICE_BEFORE_SHIPPING = 13.78;
-  const TOTAL_TAX_PRICE_BEFORE_SHIPPING = 1.22;
+  const SHIPPING_GROSS_PRICE = 75.46;
+  const SHIPPING_NET_PRICE = 69.31;
+  const SHIPPING_TAX_PRICE = 6.15;
 
-  const TOTAL_GROSS_SHIPPING_PRICE = 69.31;
-  const TOTAL_NET_SHIPPING_PRICE = 63.65;
-  const TOTAL_TAX_SHIPPING_PRICE = 5.66;
+  const TOTAL_GROSS_PRICE_AFTER_SHIPPING = 401.54;
+  const TOTAL_NET_PRICE_AFTER_SHIPPING = 368.81;
+  const TOTAL_TAX_PRICE_AFTER_SHIPPING = 32.73;
 
-  const TOTAL_GROSS_PRICE_AFTER_SHIPPING = 84.31;
-  const TOTAL_NET_PRICE_AFTER_SHIPPING = 77.43;
-  const TOTAL_TAX_PRICE_AFTER_SHIPPING = 6.88;
+  const PRODUCT_TOTAL_GROSS_PRICE_AFTER_PROMOTION = 489.12;
+  const PRODUCT_TOTAL_NET_PRICE_AFTER_PROMOTION = 449.25;
+  const PRODUCT_TOTAL_TAX_PRICE_AFTER_PROMOTION = 39.87;
 
-  const TOTAL_GROSS_PRICE_AFTER_SHIPPING_AFTER_DISCOUNT = 75.88;
-  const TOTAL_NET_PRICE_AFTER_SHIPPING_AFTER_DISCOUNT = 69.68;
-  const TOTAL_TAX_PRICE_AFTER_SHIPPING_AFTER_DISCOUNT = 6.2;
+  const TOTAL_GROSS_PRICE_INCLUDING_ORDER_PROMOTION = 564.58;
+  const TOTAL_NET_PRICE_INCLUDING_ORDER_PROMOTION = 518.56;
+  const TOTAL_TAX_PRICE_INCLUDING_ORDER_PROMOTION = 46.02;
 
-  const TOTAL_GROSS_SHIPPING_PRICE_AFTER_DISCOUNT = 62.38;
-  const TOTAL_NET_SHIPPING_PRICE_AFTER_DISCOUNT = 57.29;
-  const TOTAL_TAX_SHIPPING_PRICE_AFTER_DISCOUNT = 5.09;
+  const UNDISCOUNTE_TOTAL_GROSS_PRICE = 727.62;
+  const UNDISCOUNTE_TOTAL_NET_PRICE = 668.31;
+  const UNDISCOUNTE_TOTAL_TAX_PRICE = 59.31;
 
   it("creates token for staff user", async () => {
     await testCase
@@ -61,14 +64,14 @@ describe("App should calculate taxes for draft order with manual total discount 
       .stores("StaffUserToken", "data.tokenCreate.token")
       .retry();
   });
-  it("creates order in channel pricesEnteredWithTax: True", async () => {
+  it("creates order in channel pricesEnteredWithTax: False", async () => {
     await testCase
       .step("Create order in channel")
       .spec()
       .post("/graphql/")
       .withGraphQLQuery(CreateDraftOrder)
       .withGraphQLVariables({
-        "@DATA:TEMPLATE@": "DraftOrder:PricesWithTax",
+        "@DATA:TEMPLATE@": "DraftOrder:PricesWithoutTax",
       })
       .withHeaders({
         Authorization: "Bearer $S{StaffUserToken}",
@@ -85,29 +88,35 @@ describe("App should calculate taxes for draft order with manual total discount 
       })
       .stores("OrderID", "data.draftOrderCreate.order.id");
   });
-  it("should create order lines as staff user", async () => {
+  it("should add lines to draft order", async () => {
     await testCase
-      .step("Create order lines as staff user")
+      .step("add prodcuts")
       .spec()
       .post("/graphql/")
       .withGraphQLQuery(CreateOrderLines)
       .withGraphQLVariables({
         orderId: "$S{OrderID}",
-        input: [{ quantity: 10, variantId: "$M{Product.Juice.variantId}" }],
+        input: [
+          {
+            quantity: 1,
+            variantId: "$M{Product.ExpensiveTshirt.variantId}",
+          },
+        ],
       })
       .withHeaders({
         Authorization: "Bearer $S{StaffUserToken}",
       })
       .expectStatus(200)
-      .expectJson("data.orderLinesCreate.orderLines[0].quantity", 10)
+      .expectJson("data.orderLinesCreate.orderLines[0].quantity", 1)
       .expectJson(
-        "data.orderLinesCreate.order.total.gross.amount",
-        TOTAL_GROSS_PRICE_BEFORE_SHIPPING,
-      );
+        "data.orderLinesCreate.order.total.gross",
+        getMoney(TOTAL_NET_PRICE_BEFORE_SHIPPING, CURRENCY),
+      )
+      .stores("OrderLineId", "data.orderLinesCreate.orderLines[0].id");
   });
-  it("should update order as staff user", async () => {
+  it("should update order address", async () => {
     await testCase
-      .step("Update order as staff user")
+      .step("Update addresses on draft order")
       .spec()
       .post("/graphql/")
       .withGraphQLQuery(DraftOrderUpdateAddress)
@@ -132,9 +141,10 @@ describe("App should calculate taxes for draft order with manual total discount 
         }),
       );
   });
-  it("should update order's shipping method as staff user", async () => {
+
+  it("should update order's shipping method", async () => {
     await testCase
-      .step("Update shipping method as staff user")
+      .step("Add shipping method")
       .spec()
       .post("/graphql/")
       .withGraphQLQuery(DraftOrderUpdateShippingMethod)
@@ -161,52 +171,79 @@ describe("App should calculate taxes for draft order with manual total discount 
       .expectJson(
         "data.orderUpdateShipping.order.shippingPrice",
         getCompleteMoney({
-          gross: TOTAL_GROSS_SHIPPING_PRICE,
-          net: TOTAL_NET_SHIPPING_PRICE,
-          tax: TOTAL_TAX_SHIPPING_PRICE,
+          gross: SHIPPING_GROSS_PRICE,
+          net: SHIPPING_NET_PRICE,
+          tax: SHIPPING_TAX_PRICE,
+          currency: CURRENCY,
+        }),
+      )
+      .expectJson(
+        "data.orderUpdateShipping.order.undiscountedTotal",
+        getCompleteMoney({
+          gross: TOTAL_GROSS_PRICE_AFTER_SHIPPING,
+          net: TOTAL_NET_PRICE_AFTER_SHIPPING,
+          tax: TOTAL_TAX_PRICE_AFTER_SHIPPING,
           currency: CURRENCY,
         }),
       );
   });
-  it("should add manual discount for total as staff user", async () => {
+
+  it("should update order line", async () => {
     await testCase
-      .step("add manual discount to draft order")
+      .step("Update order line")
       .spec()
       .post("/graphql/")
-      .withGraphQLQuery(OrderDiscountAdd)
+      .withGraphQLQuery(OrderLineUpdate)
       .withGraphQLVariables({
-        orderId: "$S{OrderID}",
+        lineId: "$S{OrderLineId}",
         input: {
-          reason: "manual discount for client",
-          value: 10,
-          valueType: "PERCENTAGE",
+          quantity: 2,
         },
       })
       .withHeaders({
         Authorization: "Bearer $S{StaffUserToken}",
       })
       .expectStatus(200)
-      .expectJson("data.orderDiscountAdd.order.id", "$S{OrderID}")
+      .expectJson("data.orderLineUpdate.order.id", "$S{OrderID}")
       .expectJson(
-        "data.orderDiscountAdd.order.total",
+        "data.orderLineUpdate.order.lines[0].totalPrice",
         getCompleteMoney({
-          gross: TOTAL_GROSS_PRICE_AFTER_SHIPPING_AFTER_DISCOUNT,
-          net: TOTAL_NET_PRICE_AFTER_SHIPPING_AFTER_DISCOUNT,
-          tax: TOTAL_TAX_PRICE_AFTER_SHIPPING_AFTER_DISCOUNT,
+          gross: PRODUCT_TOTAL_GROSS_PRICE_AFTER_PROMOTION,
+          net: PRODUCT_TOTAL_NET_PRICE_AFTER_PROMOTION,
+          tax: PRODUCT_TOTAL_TAX_PRICE_AFTER_PROMOTION,
           currency: CURRENCY,
         }),
       )
       .expectJson(
-        "data.orderDiscountAdd.order.shippingPrice",
+        "data.orderLineUpdate.order.total",
         getCompleteMoney({
-          gross: TOTAL_GROSS_SHIPPING_PRICE_AFTER_DISCOUNT,
-          net: TOTAL_NET_SHIPPING_PRICE_AFTER_DISCOUNT,
-          tax: TOTAL_TAX_SHIPPING_PRICE_AFTER_DISCOUNT,
+          gross: TOTAL_GROSS_PRICE_INCLUDING_ORDER_PROMOTION,
+          net: TOTAL_NET_PRICE_INCLUDING_ORDER_PROMOTION,
+          tax: TOTAL_TAX_PRICE_INCLUDING_ORDER_PROMOTION,
           currency: CURRENCY,
         }),
       )
-      .expectJson("data.orderDiscountAdd.order.discounts[0].type", "MANUAL");
+      .expectJson(
+        "data.orderLineUpdate.order.shippingPrice",
+        getCompleteMoney({
+          gross: SHIPPING_GROSS_PRICE,
+          net: SHIPPING_NET_PRICE,
+          tax: SHIPPING_TAX_PRICE,
+          currency: CURRENCY,
+        }),
+      )
+      .expectJson(
+        "data.orderLineUpdate.order.undiscountedTotal",
+        getCompleteMoney({
+          gross: UNDISCOUNTE_TOTAL_GROSS_PRICE,
+          net: UNDISCOUNTE_TOTAL_NET_PRICE,
+          tax: UNDISCOUNTE_TOTAL_TAX_PRICE,
+          currency: CURRENCY,
+        }),
+      )
+      .expectJson("data.orderLineUpdate.order.discounts[0].type", "ORDER_PROMOTION");
   });
+
   it("should complete draft order", async () => {
     await testCase
       .step("Complete draft order")
