+++ apps/avatax/src/env.ts
@@ -25,6 +25,7 @@ export const env = createEnv({
     DYNAMODB_LOGS_TABLE_NAME: z.string(),
     E2E_USER_NAME: z.string().optional(),
     E2E_USER_PASSWORD: z.string().optional(),
+    E2E_SALEOR_VERSION: z.enum(["319", "320", "321", "latest"]).optional(),
     FILE_APL_PATH: z.string().optional(),
     MANIFEST_APP_ID: z.string().optional().default("saleor.app.avatax"),
     OTEL_ENABLED: booleanSchema.optional().default("false"),
@@ -72,6 +73,7 @@ export const env = createEnv({
     DYNAMODB_LOGS_TABLE_NAME: process.env.DYNAMODB_LOGS_TABLE_NAME,
     E2E_USER_NAME: process.env.E2E_USER_NAME,
     E2E_USER_PASSWORD: process.env.E2E_USER_PASSWORD,
+    E2E_SALEOR_VERSION: process.env.E2E_SALEOR_VERSION,
     ENV: process.env.ENV,
     FILE_APL_PATH: process.env.FILE_APL_PATH,
     MANIFEST_APP_ID: process.env.MANIFEST_APP_ID,
